Updated lock to use long strength.

+++ Whetstone.Core.Tasks/Lock.cs
@@ -45,7 +45,7 @@ namespace Whetstone.Core.Tasks
         }
 
         long FOwner = TaskContext.C_NoId;
-        int FStrength;
+        long FStrength;
 
         [NotNull]
         readonly Queue FQueue = new Queue();
